Updated the project to use roc-config + runners
This commit continues the refactor and uses roc-config for a lot of things.
Also have runners now to make building and watching the application easier, moving responsibility from the CLI to the project.
The progressbar for the builder now tries to fit inside the window.

Other improvements and fixes throughout the project.